+++ .circleci/config.yml
@@ -1,9 +1,9 @@
 version: 2
 jobs:
   build:
-    working_directory: ~/Headstart
+    working_directory: ~/headstart
     docker:
-      - image: circleci/node:6.9.0
+      - image: circleci/node:6.10.3
       - image: mongo:3.4.4
     steps:
       - checkout
@@ -22,14 +22,3 @@ jobs:
       - run:
           name: test
           command: npm test
-      - run:
-          name: code-coverage
-          command: './node_modules/.bin/nyc report --reporter=text-lcov'
-      - store_artifacts:
-          path: test-results.xml
-          prefix: tests
-      - store_artifacts:
-          path: coverage
-          prefix: coverage
-      - store_test_results:
-          path: test-results.xml
