--- config/database.js
@@ -1,1 +1,0 @@
-module.exports = { 'url': process.env.DB_URL }
