+++ views/pages/error.ejs
@@ -1,0 +1,23 @@
+<!doctype html>
+<html>
+<head>
+    <% include ../partials/head %>
+</head>
+<body>
+<div class="container">
+    <div class="page-header text-center">
+        <h1> 404 </h1>
+        <h2>Looks like you've found something that is bad :O</h2>
+        <a href="/logout" class="btn btn-default btn-sm">Logout</a>
+    </div>
+     <div class="page-header text-center">
+        <h3> Click <a href='/home'>here</a> to return home!! </h3>
+    </div>
+
+</div>
+</body>
+
+<footer>
+    <% include ../partials/footer %>
+</footer>
+</html>
