+++ views/pages/login.ejs
@@ -13,7 +13,7 @@
             <h1 id="loginThemeTitle"></h1>
 
             <!-- April 10th: Adding this now -->
-            <h2 id="buRegistrationMessage">Note: Registration is now only open to local Boston students. Global registration has closed.</h2>
+            <!-- <h2 id="buRegistrationMessage">Note: Registration is now only open to local Boston students. Global registration has closed.</h2> -->
 
             <!-- Server Messages -->
             <div class="text-center">
@@ -64,7 +64,3 @@
 
 </footer>
 </html>
-
-
-
-
